fix: correct checkpoint tests after CI
Move the composition test to NEURON: NEST cannot compose two simulations
in one adapter (its kernel locks the time resolution once nodes exist), so
the composition+streaming path is only exercisable on NEURON, where it is
already a supported path.

Fix the same-name rerun test: the checkpoint controller keeps its step
counter across run_simulation calls, so the second run does not reproduce
the first run's segment count. Assert instead that the first run's block
survives the rerun intact (no overwrite) and the second run writes its own
block.
